google-sso-fastapi/requirements.txt
fastapi
uvicorn[standard]
authlib
python-dotenv
